Add files via upload
I came up with a successful way to obtain MODIS/USDM comparisons for full years but didn’t finish it since I want to stop for tonight and also I need to deal with the two bad files (10/16/13 and 11/16/24). I stopped after April 22nd and will add the rest of the dates to the function later. Plus, I still need to individually download all the ESI files. I’ll probably just remove Nov 16 from the function when getting 2024 data and then add it back into the function.

Still, this is a very interesting discovery. I did this by finding code on stack exchange that rounds to the nearest Tuesday, used that to get the requisite USDM file for a specific date, and since the MODIS data largely uses the same dates for every year (I had to make 2 separate functions, one for leap years and one for non leap years), this should make getting MODIS/USDM merges a breeze once I actually get the MODIS ESI data for individual dates

I confirm that 2024-01-09 is the same for both this document and the old one, therefore my new algorithm is working properly.